Stop duplicate Clients.

+++ src/UI/RPC/Modules/NetworkModule.nim
@@ -17,18 +17,18 @@ import strutils
 import json
 
 #Default network port.
-const DEFAULT_PORT {.intdefine.}: int = 5132
+const DEFAULT_PORT {.intdefine.}: uint = 5132
 
 #Connect to a new node.
 proc connect*(
     rpc: RPC,
     ip: string,
-    port: int
+    port: uint
 ): Future[JSONNode] {.async.} =
     try:
         #Connect to a new node.
         if not await rpc.events.get(
-            proc (ip: string, port: int): Future[bool],
+            proc (ip: string, port: uint): Future[bool],
             "network.connect"
         )(ip, port):
             result = %* {
@@ -53,7 +53,7 @@ proc networkModule*(
             of "connect":
                 res = await rpc.connect(
                     json["args"][0].getStr(),
-                    if json["args"].len == 2: json["args"][1].getInt() else: DEFAULT_PORT
+                    if json["args"].len == 2: uint(json["args"][1].getInt()) else: DEFAULT_PORT
                 )
 
             else:
